refactor(cli): move `agent pair` out of agent.rs, which crossed the cap
commands/agent.rs reached 502 lines. The interactive pairing command is the
natural seam: it is self-contained, and it is now one of two directions the
ceremony can run in, so it reads better named than buried.

Exact piecewise copy, no behaviour change. The module header records why both
directions exist — this one is the stronger, because nothing has to carry a
secret to another machine and so nothing lands in a shell history; the inverted
one exists for the case it cannot serve, a headless box with no screen to read
a code from. It also records that this binds the peer port itself and is
therefore for a machine whose agent is not already running.
